requirements.txt
opencv-contrib-python
numpy
ev3-dc
picamera2
pyyaml